refactor: taggedlen && full rawsmallvec
